refactor: reorganize and clean code
