+++ packages/liferay-npm-bundler/src/steps/adapt/angular-cli/index.ts
@@ -4,8 +4,8 @@
  */
 
 import {project} from '../../../globals';
-import * as log from '../../../log';
 import adaptStaticURLsAtRuntime from '../../../transform/js/operation/adaptStaticURLsAtRuntime';
+import * as log from '../../../util/log';
 import {runPkgJsonScript} from '../../../util/run';
 import {
 	WebpackBundles,
