+++ packages/liferay-npm-bundler/src/steps/adapt/create-react-app/index.ts
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: LGPL-3.0-or-later
  */
 
-import * as log from '../../../log';
 import adaptStaticURLsAtRuntime from '../../../transform/js/operation/adaptStaticURLsAtRuntime';
+import * as log from '../../../util/log';
 import {runPkgJsonScript} from '../../../util/run';
 import {
 	WebpackBundles,
