+++ packages/liferay-npm-bundler/src/steps/adapt/index.ts
@@ -17,11 +17,11 @@ import {
 import path from 'path';
 
 import {bundlerGeneratedDir, project} from '../../globals';
-import * as log from '../../log';
 import exportModuleAsFunction from '../../transform/js/operation/exportModuleAsFunction';
 import namespaceWepbackJsonp from '../../transform/js/operation/namespaceWepbackJsonp';
 import replace from '../../transform/text/operation/replace';
 import {copyFiles, findFiles} from '../../util/files';
+import * as log from '../../util/log';
 import Renderer from '../../util/renderer';
 
 /**
