+++ packages/liferay-npm-bundler/src/steps/bundle/write-results.ts
@@ -7,7 +7,7 @@ import fs from 'fs-extra';
 import webpack from 'webpack';
 
 import {bundlerWebpackDir} from '../../globals';
-import * as log from '../../log';
+import * as log from '../../util/log';
 
 export default function writeResults(stats: webpack.Stats): void {
 	const {compilation} = stats;
